refactor: Rewrite BaseDialogFragment and ConfirmOptionDialogFragment; Confirmation for logout action in ProfileFragment
1. Dialogs constructed using AlertDialog's Builder in BaseDialogFragment are treated as Template Dialogs where only the required elements of the Dialog to be templated are initialized. This allows for templated dialogs like ConfirmOptionDialogFragment to have fewer/needed elements set at runtime, leading to other elements that were initialized only in the templated version (ConfirmOptionDialogFragment) to be hidden in the custom (runtime) version. In short, dialogs are built with required elements dynamically and lazily on the fly.
2. In order to initialize only the required elements of templated dialogs, BaseDialogMetadata needs to be subclassed (like ConfirmOptionDialogMetadata) and must be provided with an array of Key-Value Pairs for the elements that are required to be initialized in your custom (runtime) version of the templated dialog (like ConfirmOptionDialogFragment).
3. AlertDialogExt has extension functions on AlertDialog to help with hiding the elements of the templated dialog which are not required.
4. ProfileFragment and ProfileViewModel are rewritten to show a confirmation dialog for Logout action.
5. PostDetailActivity and PostDetailViewModel are rewritten to use ConfirmOptionDialogMetadata with ConfirmOptionDialogFragment for delete post confirmation.

+++ app/src/main/java/com/mindorks/kaushiknsanji/instagram/demo/ui/detail/PostDetailActivity.kt
@@ -16,6 +16,7 @@ import com.mindorks.kaushiknsanji.instagram.demo.databinding.ActivityPostDetailB
 import com.mindorks.kaushiknsanji.instagram.demo.di.component.ActivityComponent
 import com.mindorks.kaushiknsanji.instagram.demo.ui.base.BaseActivity
 import com.mindorks.kaushiknsanji.instagram.demo.ui.common.dialogs.option.ConfirmOptionDialogFragment
+import com.mindorks.kaushiknsanji.instagram.demo.ui.common.dialogs.option.ConfirmOptionDialogMetadata
 import com.mindorks.kaushiknsanji.instagram.demo.ui.common.dialogs.option.SharedConfirmOptionDialogViewModel
 import com.mindorks.kaushiknsanji.instagram.demo.ui.common.dialogs.progress.ProgressTextDialogFragment
 import com.mindorks.kaushiknsanji.instagram.demo.ui.common.dialogs.progress.SharedProgressTextViewModel
@@ -255,20 +256,23 @@ class PostDetailActivity : BaseActivity<PostDetailViewModel>() {
         }
 
         // Register an observer on Delete Post confirmation Request events
-        viewModel.launchDeletePostConfirm.observeResourceEvent(this) { _, titleResId: Int ->
+        viewModel.launchDeletePostConfirm.observeEvent(this) { metadata: ConfirmOptionDialogMetadata ->
             // Show the Confirmation Dialog to the User for Delete Post request
             dialogManager.showDialog(
                 ConfirmOptionDialogFragment::class.java,
                 ConfirmOptionDialogFragment.Companion::newInstance
             )
-            // Pass the Dialog Title to be shown via its Shared ViewModel
-            sharedConfirmOptionDialogViewModel.onDialogTitleTextChange(Resource.Success(titleResId))
+            // Pass the metadata of the Dialog to be shown via its Shared ViewModel
+            sharedConfirmOptionDialogViewModel.onDialogMetadataChange(metadata)
         }
 
         // Register an observer on Delete Post confirmation - Positive response events
         sharedConfirmOptionDialogViewModel.actionPositiveButton.observeEvent(this) {
-            // Delegate to the Primary ViewModel to begin the delete process
-            viewModel.onDeleteConfirm()
+            // Check if the Dialog Confirmation response is for DeletePost Dialog type
+            if (sharedConfirmOptionDialogViewModel.isDialogType(PostDetailViewModel.CONFIRM_OPTION_DIALOG_TYPE_DELETE_POST)) {
+                // Delegate to the Primary ViewModel to begin the delete process
+                viewModel.onDeleteConfirm()
+            }
         }
 
         // Register an observer to return to the Calling Activity with the status of User's Like on the Post
